Tidy up wallet UI
