+++ sdk/typescript/docs/pages/playground/cosmos-kit.mdx
@@ -3,6 +3,7 @@ import {CosmosKit} from "../../components/cosmos-kit";
 
 # Cosmos Kit
 
-Below is an example that uses [CosmosKit](https://cosmoskit.com/) to connect your [Keplr wallet](https://www.keplr.app/) to this page:
+Below is an example that uses [CosmosKit](https://cosmoskit.com/) to connect your [Keplr wallet](https://www.keplr.app/) or
+[Ledger hardware wallet](https://www.ledger.com/) to this page:
 
 <CosmosKit />
